멀티 모듈로 전환
