+++ application/build.gradle
@@ -1,0 +1,9 @@
+jar.enabled = false
+bootJar.enabled = true
+
+dependencies {
+    implementation(project(":domain"))
+    implementation(project(":web"))
+    implementation 'org.springframework.boot:spring-boot-starter'
+}
+
