+++ domain/build.gradle
@@ -1,0 +1,9 @@
+jar.enabled = true
+bootJar.enabled = false
+
+dependencies {
+    implementation(project(":library"))
+    implementation 'org.springframework:spring-context'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.15.2'
+}
