+++ gradle.properties
@@ -1,0 +1,8 @@
+### Application version ###
+applicationVersion=0.0.1-SNAPSHOT
+
+### Project configs ###
+projectGroup=com.weather
+javaVersion=17
+springBootVersion=3.3.3
+springDependencyManagement=1.1.6
