+++ library/build.gradle
@@ -1,0 +1,7 @@
+jar.enabled = true
+bootJar.enabled = false
+
+dependencies {
+    implementation(project(":spring"))
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
