+++ settings.gradle
@@ -1,1 +1,20 @@
+pluginManagement {
+
+    repositories {
+        gradlePluginPortal()
+        mavenCentral()
+    }
+    plugins {
+        id 'org.springframework.boot' version '3.3.3'
+        id 'io.spring.dependency-management' version '1.1.6'
+    }
+}
+
 rootProject.name = 'weather-fcst'
+include 'domain'
+include 'application'
+include 'spring'
+include 'web'
+include 'library'
+include 'java'
+include 'test'
