+++ spring/build.gradle
@@ -1,0 +1,8 @@
+jar.enabled = true
+bootJar.enabled = false
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-aop'
+    implementation 'org.aspectj:aspectjweaver:1.9.19'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
