+++ spring/src/main/java/com/weather/weatherfcst/config/RestTemplateConfig.java
@@ -1,0 +1,17 @@
+package com.weather.weatherfcst.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.converter.xml.MappingJackson2XmlHttpMessageConverter;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        RestTemplate restTemplate = new RestTemplate();
+        restTemplate.getMessageConverters().add(new MappingJackson2XmlHttpMessageConverter());
+        return restTemplate;
+    }
+}
