--- src/main/java/com/weather/weatherfcst/common/RestTemplateConfig.java
@@ -1,17 +1,0 @@
-package com.weather.weatherfcst.common;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.converter.xml.MappingJackson2XmlHttpMessageConverter;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestTemplateConfig {
-
-    @Bean
-    public RestTemplate restTemplate() {
-        RestTemplate restTemplate = new RestTemplate();
-        restTemplate.getMessageConverters().add(new MappingJackson2XmlHttpMessageConverter());
-        return restTemplate;
-    }
-}
