--- src/test/java/com/weather/weatherfcst/WeatherFcstApplicationTests.java
@@ -1,13 +1,0 @@
-package com.weather.weatherfcst;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WeatherFcstApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
