+++ test/build.gradle
@@ -1,0 +1,7 @@
+jar.enabled = true
+bootJar.enabled = false
+
+dependencies {
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+}
